Split statements into multiple database commands:
issuing multiple sql statements in the same command is a undocumented feature, probably exploited for performance reason, but not officially supported. The dml api will check if the first statement (supposed to be the only one) is actually a DML one. As the first statement in the chain is a select, the command fails.

Fixes xolp automated tests